rename utils.js to helpers.js and move to /utils
